+++ CHANGELOG.md
@@ -38,8 +38,10 @@
 
 ### ❌ Removals
 - Measures: Remove effect size - Log-frequency biased MD / Mutual Dependency
+- Utils: Remove NLTK's Malayalam Punkt sentence tokenizer
 
 ### ⏫ Dependency Changes
+- Dependencies: Upgrade NLTK to 3.9.1
 - Dependencies: Upgrade Pyphen to 0.17.0
 - Dependencies: Upgrade python-mecab-ko to 1.3.7
 - Dependencies: Upgrade python-pptx to 1.0.2